Architecture cleanup
