+++ src/main/java/cuziArchitecture/pages/LandingPage.java
@@ -1,4 +1,4 @@
-package cuziArchitecture.driver;
+package cuziArchitecture.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
