+++ src/test/java/cuziArchitecture/misctests/DayThree.java
@@ -1,4 +1,4 @@
-package CuziTestArchitecture.browsertests;
+package cuziArchitecture.misctests;
 
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Parameters;
